cs336-data/pyproject.toml
[project]
name = "assignment4-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasketch>=1.7.0",
    "fastparquet>=2024.11.0",
    "fasttext-wheel>=0.9.2",
    "numpy<2",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pytest>=9.0.1",
    "requests>=2.32.5",
    "resiliparse>=0.15.2",
    "warcio>=1.7.5",
    "xopen>=2.0.2",
]

[tool.uv.workspace]
members = [
    "cs336-basics",
]
